[update] rename struct
